move surface back to classes, add methods to field class
